refactor(wallet): isolate masternode operator implementation
