+++ test/util/data/non-backported.txt
@@ -87,5 +87,6 @@ src/wallet/bip39*
 src/wallet/coinjoin.*
 src/wallet/hdchain.*
 src/wallet/masternode_operator.*
+src/wallet/mnoperator.cpp
 src/wallet/test/masternode_operator_tests.cpp
 src/hash_x11.h
